requirements.txt
pywinpty
sounddevice
tts
torch
torchaudio --index-url https://download.pytorch.org/whl/cu121
openai-whisper